+++ Specs/Braintree-API-Integration-Specs/BTClientCoinbaseIntegrationSpec.m
@@ -14,13 +14,11 @@ beforeEach(^{
                                             BTClientTestConfigurationKeyClientTokenVersion: @2
                                             } async:@YES completion:^(BTClient *client) {
                                                 testClient = client;
-                                                
-
                                                 [testClient updateCoinbaseMerchantOptions:@{ @"enabled": @YES }
                                                                                   success:^{
                                                                                       [fetchTestClientTokenExpectation fulfill];
                                                                                   } failure:^(NSError *error) {
-                                                                                      XCTFail(@"HI");
+                                                                                      XCTFail(@"Should not receive call to failure block");
                                                                                   }];
                                             }];
     [self waitForExpectationsWithTimeout:10 handler:nil];
@@ -35,11 +33,12 @@ afterEach(^{
     [self waitForExpectationsWithTimeout:10 handler:nil];
 });
 
-describe(@"saveCoinbaseAccount:success:failure:", ^{
+describe(@"saveCoinbaseAccount:storeInVault:success:failure:", ^{
     it(@"returns a valid response including a nonce in the happy path", ^{
         XCTestExpectation *tokenizationExpectation = [self expectationWithDescription:@"Tokenize Coinbase account"];
         id coinbaseResponse = @{ @"code": @"fake_coinbase_auth_code" };
         [testClient saveCoinbaseAccount:coinbaseResponse
+                           storeInVault:NO
                                 success:^(BTCoinbasePaymentMethod *coinbase){
                                     XCTAssertTrue([coinbase isKindOfClass:[BTCoinbasePaymentMethod class]]);
                                     expect(coinbase.nonce).to.beANonce();
@@ -57,6 +56,7 @@ describe(@"saveCoinbaseAccount:success:failure:", ^{
         XCTestExpectation *tokenizationExpectation = [self expectationWithDescription:@"Tokenize invalid auth code"];
         id coinbaseResponse = @{ @"code": @"fake_coinbase_invalid_authorization_code" };
         [testClient saveCoinbaseAccount:coinbaseResponse
+                           storeInVault:NO
                                 success:^(BTCoinbasePaymentMethod *coinbase){
                                     XCTFail(@"Should not receive call to Coinbase tokenization success block");
                                 } failure:^(NSError *error){
@@ -73,6 +73,7 @@ describe(@"saveCoinbaseAccount:success:failure:", ^{
         XCTestExpectation *tokenizationExpectation = [self expectationWithDescription:@"Tokenize malformed coinbase response"];
         id coinbaseResponse = @"this is not a dictionary containing an auth code";
         [testClient saveCoinbaseAccount:coinbaseResponse
+                           storeInVault:NO
                                 success:^(BTCoinbasePaymentMethod *coinbase){
                                     XCTFail(@"Should not receive call to Coinbase tokenization success block");
                                 } failure:^(NSError *error){
@@ -106,6 +107,7 @@ describe(@"saveCoinbaseAccount:success:failure:", ^{
         
         id coinbaseResponse = @{ @"code": @"1234" };
         [testClient saveCoinbaseAccount:coinbaseResponse
+                           storeInVault:NO
                                 success:nil
                                 failure:^(NSError *error){
                                     XCTFail(@"Should not receive call to Coinbase tokenization failure block");
@@ -133,12 +135,108 @@ describe(@"saveCoinbaseAccount:success:failure:", ^{
         
         id coinbaseResponse = @{ @"code": @"1234" };
         [testClient saveCoinbaseAccount:coinbaseResponse
+                           storeInVault:NO
                                 success:^(BTCoinbasePaymentMethod *coinbase){
                                     XCTFail(@"Should not receive call to Coinbase tokenization success block");
                                 } failure:nil];
         
         [mockClient verifyWithDelay:10];
         [self waitForExpectationsWithTimeout:10 handler:nil];
+    });
+
+    describe(@"storeInVault", ^{
+        it(@"stores the coinbase account in the customer's vault when YES", ^{
+            XCTestExpectation *tokenizationExpectation = [self expectationWithDescription:@"Tokenize Coinbase account"];
+            id coinbaseResponse = @{ @"code": @"fake_coinbase_auth_code" };
+            [testClient saveCoinbaseAccount:coinbaseResponse
+                               storeInVault:YES
+                                    success:^(BTCoinbasePaymentMethod *coinbase){
+                                        XCTAssertTrue([coinbase isKindOfClass:[BTCoinbasePaymentMethod class]]);
+                                        [tokenizationExpectation fulfill];
+                                    } failure:^(NSError *error){
+                                        XCTFail(@"Should not receive call to Coinbase tokenization failure block");
+                                    }];
+            
+            [self waitForExpectationsWithTimeout:10 handler:nil];
+            
+            XCTestExpectation *fetchPaymentMethodsExpectation = [self expectationWithDescription:@"Fetch Payment Methods"];
+            [testClient fetchPaymentMethodsWithSuccess:^(NSArray *paymentMethods) {
+                expect(paymentMethods).to.haveCountOf(1);
+                expect(paymentMethods[0]).to.beKindOf([BTCoinbasePaymentMethod class]);
+                [fetchPaymentMethodsExpectation fulfill];
+            } failure:^(NSError *error) {
+                XCTFail(@"Should not receive call to Fetch payment methods tokenization failure block");
+            }];
+            
+            [self waitForExpectationsWithTimeout:10 handler:nil];
+        });
+
+        it(@"does not store the coinbase account in the customer's vault when NO", ^{
+            XCTestExpectation *tokenizationExpectation = [self expectationWithDescription:@"Tokenize Coinbase account"];
+            id coinbaseResponse = @{ @"code": @"fake_coinbase_auth_code" };
+            [testClient saveCoinbaseAccount:coinbaseResponse
+                               storeInVault:NO
+                                    success:^(BTCoinbasePaymentMethod *coinbase){
+                                        XCTAssertTrue([coinbase isKindOfClass:[BTCoinbasePaymentMethod class]]);
+                                        [tokenizationExpectation fulfill];
+                                    } failure:^(NSError *error){
+                                        XCTFail(@"Should not receive call to Coinbase tokenization failure block");
+                                    }];
+            
+            [self waitForExpectationsWithTimeout:10 handler:nil];
+
+            XCTestExpectation *fetchPaymentMethodsExpectation = [self expectationWithDescription:@"Fetch Payment Methods"];
+            [testClient fetchPaymentMethodsWithSuccess:^(NSArray *paymentMethods) {
+                expect(paymentMethods).to.haveCountOf(0);
+                [fetchPaymentMethodsExpectation fulfill];
+            } failure:^(NSError *error) {
+                XCTFail(@"Should not receive call to Fetch payment methods tokenization failure block");
+            }];
+
+            [self waitForExpectationsWithTimeout:10 handler:nil];
+        });
+
+        it(@"has no impact when no customer is present", ^{
+            XCTestExpectation *fetchTestClientExpectation = [self expectationWithDescription:@"Fetch test client"];
+            [BTClient testClientWithConfiguration:@{
+                                                    BTClientTestConfigurationKeyMerchantIdentifier:@"integration_merchant_id",
+                                                    BTClientTestConfigurationKeyPublicKey:@"integration_public_key",
+                                                    BTClientTestConfigurationKeyNoCustomer: @YES,
+                                                    BTClientTestConfigurationKeyClientTokenVersion: @2
+                                                    } async:@YES completion:^(BTClient *client) {
+                                                        testClient = client;
+                                                        [testClient updateCoinbaseMerchantOptions:@{ @"enabled": @YES }
+                                                                                          success:^{
+                                                                                              [fetchTestClientExpectation fulfill];
+                                                                                          } failure:^(NSError *error) {
+                                                                                              XCTFail(@"Should not receive call to failure block");
+                                                                                          }];
+                                                    }];
+            [self waitForExpectationsWithTimeout:10 handler:nil];
+            
+            XCTestExpectation *tokenizationExpectation = [self expectationWithDescription:@"Tokenize Coinbase account"];
+            id coinbaseResponse = @{ @"code": @"fake_coinbase_auth_code" };
+            [testClient saveCoinbaseAccount:coinbaseResponse
+                               storeInVault:YES
+                                    success:^(BTCoinbasePaymentMethod *coinbase){
+                                        XCTAssertTrue([coinbase isKindOfClass:[BTCoinbasePaymentMethod class]]);
+                                        [tokenizationExpectation fulfill];
+                                    } failure:^(NSError *error){
+                                        XCTFail(@"Should not receive call to Coinbase tokenization failure block");
+                                    }];
+            
+            [self waitForExpectationsWithTimeout:10 handler:nil];
+            
+            XCTestExpectation *fetchPaymentMethodsExpectation = [self expectationWithDescription:@"Fetch Payment Methods"];
+            [testClient fetchPaymentMethodsWithSuccess:^(NSArray *paymentMethods) {
+                expect(paymentMethods).to.haveCountOf(0);
+                [fetchPaymentMethodsExpectation fulfill];
+            } failure:^(NSError *error) {
+                XCTFail(@"Should not receive call to Fetch payment methods tokenization failure block");
+            }];
+            
+            [self waitForExpectationsWithTimeout:10 handler:nil];
+        });
     });
 });